src/api/requirements.txt
alembic==1.11.1
fastapi==0.103.0
pydantic[email]
datetime==5.2
python-jose[cryptography]
passlib[bcrypt]
python-multipart==0.0.6
psycopg2-binary==2.9.6
sqlmodel==0.0.8
uvicorn==0.22.0
boto3==1.28.36
#pydantic_settings==2.0.1